requirements.txt
docx2pdf
PyYAML
tqdm
rich
python-dotenv
